+++ bukkit/src/main/java/de/maxhenkel/voicechat/voice/server/ServerVoiceEvents.java
@@ -79,10 +79,10 @@ public class ServerVoiceEvents implements Listener {
     public static void sendIncompatibleMessage(Player player, int clientCompatibilityVersion) {
         if (clientCompatibilityVersion <= 6) {
             // Send a literal string, as we don't know if the translations exist on these versions
-            player.sendMessage(String.format(Voicechat.TRANSLATIONS.voicechatNotCompatibleMessage.get(), Voicechat.INSTANCE.getDescription().getVersion(), BuildConstants.PLUGIN_NAME));
+            player.sendMessage(String.format(Voicechat.TRANSLATIONS.voicechatNotCompatibleMessage.get(), BuildConstants.MOD_COMPATIBLE_VERSION, BuildConstants.PLUGIN_NAME));
         } else {
             // This translation key is only available for compatibility version 7+
-            Voicechat.compatibility.sendIncompatibleMessage(player, Voicechat.INSTANCE.getDescription().getVersion(), BuildConstants.PLUGIN_NAME);
+            Voicechat.compatibility.sendIncompatibleMessage(player, BuildConstants.MOD_COMPATIBLE_VERSION, BuildConstants.PLUGIN_NAME);
         }
     }