+++ forge/changelog.md
@@ -1,1 +1,2 @@
 - Added more detailed microphone error messages when using the java microphone implementation
+- Improved incompatible version message